prestige system now absolutely complete

+++ script/engine.js
@@ -10,6 +10,7 @@ var Engine = {
 	VERSION: 1.3,
 	MAX_STORE: 99999999999999,
 	SAVE_DISPLAY: 30 * 1000,
+  GAME_OVER: false,
 	
 	//object event types
 	topics: {},
@@ -206,12 +207,12 @@ var Engine = {
 	},
 	
 	deleteSave: function() {
-    //var carriedPrestige = $SM.get('previous.stores');
-		if(typeof Storage != 'undefined' && localStorage) {
-			localStorage.clear();
-		}
+    if (!Engine.GAME_OVER) {
+      if(typeof Storage != 'undefined' && localStorage) {
+        localStorage.clear();
+      }
+    }
 		location.reload();
-    //$SM.set('previous.stores',carriedPrestige);
 	},
 	
 	share: function() {
